Move create_temperature_dummy to helpers and integrate it in prepare_data function
